Backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
chromadb
python-multipart
requests
pydantic
cryptography
PyMuPDF
tiktoken
httpx
sentence-transformers
google-genai
transformers
torch
accelerate
huggingface-hub
supabase
python-jose[cryptography]
passlib[bcrypt]
